requirements.txt
a2wsgi==1.10.8
aiohappyeyeballs==2.6.1
aiohttp==3.8.5
aiologic==0.14.0
aiosignal==1.3.1
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.0.0
apache-airflow-core==3.0.0
apache-airflow-providers-common-compat==1.6.0
apache-airflow-providers-common-io==1.5.4
apache-airflow-providers-common-sql==1.26.0
apache-airflow-providers-ftp==3.5.1
apache-airflow-providers-http==4.5.1
apache-airflow-providers-imap==3.3.1
apache-airflow-providers-postgres==5.11.3
apache-airflow-providers-smtp==2.0.3
apache-airflow-providers-sqlite==3.4.3
apache-airflow-providers-standard==1.0.0
apache-airflow-task-sdk==1.0.0
apispec==6.3.0
argcomplete==3.1.1
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.3.0
Automat==22.10.0
Babel==2.12.1
backoff==1.10.0
bcrypt==3.2.2
blinker==1.6.2
cachelib==0.9.0
cadwyn==5.3.3
cattrs==23.1.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
clickclick==20.10.2
cloud-init==24.4.1
colorama==0.4.6
colorlog==6.9.0
command-not-found==0.3
configobj==5.0.8
ConfigUpdater==3.1.1
connexion==2.14.2
constantly==23.10.4
cron_descriptor==1.4.0
croniter==6.0.0
cryptography==41.0.3
dbus-python==1.3.2
Deprecated==1.2.14
dill==0.3.1.1
distlib==0.3.9
distro==1.9.0
distro-info==1.7+build1
dnspython==2.4.2
docutils==0.20.1
email_validator==2.2.0
Faker==37.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.5.0
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.4.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-re2==1.1
googleapis-common-protos==1.70.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.71.0
gunicorn==21.2.0
gyp==0.1
h11==0.14.0
httpcore==1.0.8
httplib2==0.20.4
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
incremental==22.10.0
inflection==0.5.1
iotop==0.6
itsdangerous==2.1.2
Jinja2==3.1.6
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
kafka-python==2.2.11
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
libcst==1.7.0
limits==3.6.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.7.0
msgspec==0.19.0
multidict==6.0.4
netifaces==0.11.0
oauthlib==3.2.2
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
ordered-set==4.1.0
packaging==25.0
pathspec==0.11.2
pendulum==2.1.2
platformdirs==4.3.7
pluggy==1.5.0
prison==0.2.1
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.25.1
pycparser==2.21
pycurl==7.45.3
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.16.1
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.10.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
pyspark==3.5.5
python-apt==2.7.7+ubuntu4
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.1.0
python-multipart==0.0.20
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
retryhttp==1.3.2
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==14.0.0
rich-argparse==1.3.0
rich-toolkit==0.14.1
rpds-py==0.10.2
service-identity==24.1.0
setproctitle==1.3.5
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.0
speedtest-cli==2.1.3
SQLAlchemy==1.4.49
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
starlette==0.46.2
structlog==25.2.0
svcs==25.1.0
systemd-python==235
tabulate==0.9.0
tenacity==8.2.3
termcolor==3.0.1
text-unidecode==1.3
Twisted==24.3.0
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ubuntu-pro-client==8001
uc-micro-py==1.0.2
unattended-upgrades==0.1
unicodecsv==0.14.1
universal_pathlib==0.2.6
urllib3==2.0.7
uuid6==2024.7.10
uvicorn==0.27.1
uvloop==0.19.0
vboxapi==1.0
virtualenv==20.30.0
wadllib==1.3.6
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.16.2
zope.interface==6.1
